# frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactVersionDetail": {
        "properties": {
          "artifact_type": {
            "enum": [
              "课件",
              "教案",
              "提纲",
              "试卷",
              "互动内容"
            ],
            "title": "Artifact Type",
            "type": "string"
          },
          "content": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "download_url": {
            "title": "Download Url",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "session_id",
          "artifact_type",
          "version",
          "label",
          "origin",
          "filename",
          "title",
          "created_at",
          "download_url"
        ],
        "title": "ArtifactVersionDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "多模态 AI 备课智能体后端。\n\nOpenAPI 是本服务接口文档的**唯一事实源**：每个端点都带 `summary`、描述、\n请求/响应示例、错误码与分组 tag；前端 TypeScript 类型由本 schema 生成，禁止手抄。\n* **界面**：Swagger UI `/docs`，ReDoc `/redoc`。\n* **分组**：系统 / 备课会话 / 知识库 / 知识图谱 / 生成物 / 冲突审核 / 题库。\n* **单用户**：固定 `user_id=\"default\"`，本服务不含登录与多租户。\n* **启动清扫**：进程重启时，上一轮被中断的「处理中」资料不会有后台任务再推进，\n  服务启动时统一标记为「失败」，避免前端无限轮询（教师可按原文重新上传）。\n* **能力切换**：对话模型、向量化、语音转写、PDF 解析、网络搜索均按配置选择。\n  云端能力未配置时相关端点返回 `503`（`code: provider_not_configured`，`message` 面向教师，\n  指到设置页）——产品不返回任何「看起来像真结果的占位内容」。\n* **生成失败**：配好了但生成没成功时返回 `502`（`code: generation_failed`），\n  教师可重试——不落一份空课件 / 空教案冒充成功。\n* **版本策略**：只有破坏性变更才升 `/api/v2`。\n",
    "title": "EduMind API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/artifacts/{version_id}": {
      "get": {
        "description": "按版本 id 取这一版的详情：版本号、产出方式（生成 / 修改）、落盘文件名、标题、基线版本（`parent_id`，即「由哪一版衍生而来」）以及**这一版的内容快照**。\n\n内容快照让历史版本可以直接回看（课件结构与页面、教案结构、提纲正文、题目、互动内容 HTML）；下载文件请用 `download_url` 指向的下载端点。\n\n版本 id 不存在时返回 `404`。",
        "operationId": "get_artifact_version_api_v1_artifacts__version_id__get",
        "parameters": [
          {
            "description": "版本 id，取自版本列表的 `versions[].id`",
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "description": "版本 id，取自版本列表的 `versions[].id`",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "artifact_type": "课件",
                  "content": {
                    "slides": [
                      {
                        "points": [
                          "情境导入：出租车计价"
                        ],
                        "role": "封面",
                        "title": "一次函数"
                      }
                    ]
                  },
                  "created_at": "2026-09-24T10:00:30",
                  "download_url": "/api/v1/artifacts/3f2a1b0c-9d8e-4f70-8a1b-2c3d4e5f6a7b/download",
                  "filename": "courseware_ppt_1a2b3c4d.pptx",
                  "id": "3f2a1b0c-9d8e-4f70-8a1b-2c3d4e5f6a7b",
                  "label": "第 2 版",
                  "origin": "修改",
                  "parent_id": "3f2a1b0c-9d8e-4f70-8a1b-2c3d4e5f6a7b",
                  "session_id": "9f1a2b3c-4d5e-4f60-8a7b-1c2d3e4f5a6b",
                  "title": "一次函数",
                  "version": 2
                },
                "schema": {
                  "$ref": "#/components/schemas/ArtifactVersionDetail"
                }
              }
            },
            "description": "某一版生成物的详情（含内容快照）"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "生成物版本不存在: 3f2a1b0c-9d8e-4f70-8a1b-2c3d4e5f6a7b"
                }
              }
            },
            "description": "生成物版本不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            },
            "description": "未捕获的服务端错误:进程存活但本次请求失败,前端应提示重试。"
          }
        },
        "summary": "取某个生成物版本的详情",
        "tags": [
          "生成物"
        ]
      }
    },
...
    "/api/v1/revise": {
      "post": {
        "description": "对**所选课件**提修改意见并重新生成：把待改的课件结构连同意见交模型重排，再渲染为**新文件**（旧文件保留，因此可反复修改互不覆盖）。\n\n本接口只作用于传入的这一份课件，不影响同次备课的教案与提纲；`style` 用于沿用原课件风格（配色主题），不传则回退默认主题。\n\n**带上 `session_id` 时本次修改落一条新版本**（`origin=修改`）：版本号比基线更高，基线版本保持可回看、可下载；`base_version_id` 指定以哪一版为基线（版本 id 取自版本列表），不传时以该会话的当前版本为基线。基线或会话不存在返回 `404`，`base_version_id` 不是课件版本返回 `422`。\n\n不带 `session_id` / `base_version_id` 时保持既有语义：只返回修改后的结构与新文件名，不落版本记录。",
        "operationId": "revise_api_v1_revise_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "filename": "courseware_ppt_6a5b4c3d.pptx",
                  "session_id": "9f1a2b3c-4d5e-4f60-8a7b-1c2d3e4f5a6b",
                  "slides": [
                    {
                      "bullets": [
                        "形如 y=kx+b（k≠0）",
                        "出租车计价就是一次函数"
                      ],
                      "title": "一次函数"
                    }
                  ],
                  "version": 2,
                  "version_id": "4a3b2c1d-0e9f-4a81-9b2c-3d4e5f6a7b8c"
                },
                "schema": {
                  "$ref": "#/components/schemas/ReviseResponse"
                }
              }
            },
            "description": "修改后的课件、新文件名与（带会话时的）新版本标识"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "生成物版本不存在: 3f2a1b0c-9d8e-4f70-8a1b-2c3d4e5f6a7b"
                }
              }
            },
            "description": "基线版本或备课会话不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "messages"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ]
                }
              }
            },
            "description": "请求校验失败:请求体、表单或路径字段缺失或类型不符"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            },
            "description": "未捕获的服务端错误:进程存活但本次请求失败,前端应提示重试。"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "provider_not_configured",
                    "message": "还没有配置模型供应商：到「设置 → 添加供应商」选一家并粘贴 Key 后即可使用。"
                  }
                }
              }
            },
            "description": "所需的云端能力未配置：不返回演示用假结果。`detail.code` 固定为 `provider_not_configured`，`detail.message` 是面向教师的一句引导（去设置页添加供应商或补齐配置）。配置后即时生效，无需重启。"
          }
        },
        "summary": "按修改意见重做课件",
        "tags": [
          "生成物"
        ]
      }
    },
    "/api/v1/revise/exam": {
      "post": {
        "description": "对**所选试卷**提修改意见并重新出题：把待改的题目集合连同意见交模型重排，渲染为**新文件**并把改后的题目写入题库（来源=`自编`，按考查知识点关联图谱节点）。\n\n本接口只作用于传入的这一份试卷，不影响同次备课的其它生成物；`n` 指定改后题量，不传时沿用基线题量。\n\n**带上 `session_id` 时本次修改落一条新版本**（`origin=修改`）：版本号比基线更高，基线版本保持可回看、可下载；`base_version_id` 指定以哪一版为基线，不传时以该会话的当前版本为基线。基线或会话不存在返回 `404`，`base_version_id` 不是试卷版本返回 `422`。\n\n不带 `session_id` / `base_version_id` 时只返回题目与新文件名，不落版本记录。",
        "operationId": "revise_exam_endpoint_api_v1_revise_exam_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseExamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "bank_saved": 5,
                  "filename": "exam_4c5d6e7f.docx",
                  "questions": [
                    {
                      "analysis": "一次函数形如 y=kx+b（k≠0）。",
                      "answer": "A",
                      "content": "下列函数中，属于一次函数的是？",
                      "knowledge_point": "一次函数的定义",
                      "options": [
                        "A. y=2x+1",
                        "B. y=x²"
                      ],
                      "type": "选择"
                    }
                  ],
                  "session_id": "9f1a2b3c-4d5e-4f60-8a7b-1c2d3e4f5a6b",
                  "version": 2,
                  "version_id": "8b7c6d5e-4f3a-42b1-8c9d-8e7f6a5b4c3d"
                },
                "schema": {
                  "$ref": "#/components/schemas/ReviseExamResponse"
                }
              }
            },
            "description": "修改后的题目、新文件名、入题库题数与（带会话时的）新版本标识"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "生成物版本不存在: 8b7c6d5e-4f3a-42b1-8c9d-8e7f6a5b4c3d"
                }
              }
            },
            "description": "基线版本或备课会话不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "messages"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ]
                }
              }
            },
            "description": "请求校验失败:请求体、表单或路径字段缺失或类型不符"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            },
            "description": "未捕获的服务端错误:进程存活但本次请求失败,前端应提示重试。"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "provider_not_configured",
                    "message": "还没有配置模型供应商：到「设置 → 添加供应商」选一家并粘贴 Key 后即可使用。"
                  }
                }
              }
            },
            "description": "所需的云端能力未配置：不返回演示用假结果。`detail.code` 固定为 `provider_not_configured`，`detail.message` 是面向教师的一句引导（去设置页添加供应商或补齐配置）。配置后即时生效，无需重启。"
          }
        },
        "summary": "按修改意见重做试卷",
        "tags": [
          "生成物"
        ]
      }
    },
    "/api/v1/revise/interactive": {
      "post": {
        "description": "对**所选互动内容**提修改意见并重新生成：把待改的单文件 HTML 连同意见交模型重排，落盘为**新文件**。\n\n结果不是完整单文件 HTML 时退回基线 HTML（不落半成品）；取回该版本请用 `GET /api/v1/artifacts/{version_id}/download?inline=true` 在新标签页打开试用。\n\n**带上 `session_id` 时本次修改落一条新版本**（`origin=修改`）：版本号比基线更高，基线版本保持可回看、可下载；`base_version_id` 指定以哪一版为基线，不传时以该会话的当前版本为基线。基线或会话不存在返回 `404`，`base_version_id` 不是互动内容版本返回 `422`。\n\n不带 `session_id` / `base_version_id` 时只返回 HTML 与新文件名，不落版本记录。",
        "operationId": "revise_interactive_endpoint_api_v1_revise_interactive_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseInteractiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "filename": "creative_5d6e7f8a.html",
                  "html": "<!doctype html><html lang=zh><body>…</body></html>",
                  "session_id": "9f1a2b3c-4d5e-4f60-8a7b-1c2d3e4f5a6b",
                  "version": 2,
                  "version_id": "9c8d7e6f-5a4b-43c2-9d0e-9f8a7b6c5d4e"
                },
                "schema": {
                  "$ref": "#/components/schemas/ReviseInteractiveResponse"
                }
              }
            },
            "description": "修改后的互动内容、新文件名与（带会话时的）新版本标识"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "生成物版本不存在: 9c8d7e6f-5a4b-43c2-9d0e-9f8a7b6c5d4e"
                }
              }
            },
            "description": "基线版本或备课会话不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "messages"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ]
                }
              }
            },
            "description": "请求校验失败:请求体、表单或路径字段缺失或类型不符"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            },
            "description": "未捕获的服务端错误:进程存活但本次请求失败,前端应提示重试。"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "provider_not_configured",
                    "message": "还没有配置模型供应商：到「设置 → 添加供应商」选一家并粘贴 Key 后即可使用。"
                  }
                }
              }
            },
            "description": "所需的云端能力未配置：不返回演示用假结果。`detail.code` 固定为 `provider_not_configured`，`detail.message` 是面向教师的一句引导（去设置页添加供应商或补齐配置）。配置后即时生效，无需重启。"
          }
        },
        "summary": "按修改意见重做互动内容",
        "tags": [
          "生成物"
        ]
      }
    },
...
  },
...
}